Taller2/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
unstructured
pypdf
